refactor(proportion_female): Remove comments
proportion_female was commented out in all of the case studies because
setting it was not doing anything to the model. In M2 we will remove the
ability to set this parameter vector because it was not doing anything.
In the future we will add the feature back. This change does nothing to
the results.

Simplified a comment in the scamp case study to say that it is fixed at
0.5 instead of you have the ability to estimate it.

Part of NOAA-FIMS/FIMS#638 and NOAA-FIMS/seaside-chats#7