Uses the wrapper functions for scamp & yellowtail
There is still something wrong with scamp. F in fleet1 spikes in 2008 and
I am not sure why.

Selectivity seems to be pretty well estimated, even for the second survey
of the yellowtail example. For this selectivity parameter, I checked the
previous version of FIMS (i.e., one commit before v0.3.0 and Rcpp 1.0.12)
and the estimate of selectivity slope for survey2 was also VERY high in
that run as well.

With scamp:
* KFJ fixed a problem with the steepness parameter not being transformed
  into logit space but I am unsure if we should be using 0.75 or 0.99
  because of a comment in the original file.

Co-authored-by: Bai-Li-NOAA <bai.li@noaa.gov>