fix(scamp): Use the wrapper functions
There is something wrong with some units or something that I did because
it does not match nearly as well as it did the previous round. Committing
so others can look at the changes and help. But, the model does converge
and the plots render. Selectivity looks to be right but fishing mortality
is on a different level in the terminal year estimate and I am not sure
why.